rename club controller/service/module with plural
