+++ api/src/clubs/clubs.module.ts
@@ -4,9 +4,9 @@ import { CaslAbilityFactory } from '../shared/modules/casl/casl-ability.factory'
 import { UsersModule } from '../users/users.module';
 import { ClubMember } from './club-member.entity';
 import { ClubSearchService } from './club-search.service';
-import { ClubsController } from './club.controller';
 import { Club } from './club.entity';
-import { ClubsService } from './club.service';
+import { ClubsController } from './clubs.controller';
+import { ClubsService } from './clubs.service';
 
 @Module({
   imports: [
